classic-java-11: Switch to zulu alpine image
This is done mainly because openjdk decided to change to a older and
more stable upstream, missing various packages we were using.

As azul now provides an Alpine image, we try this instead of trying
to get packages installed on an older debian release.

Most of this is from the normal classic build.