Fix group used in modern-v2
Use docker group for jenkins user.

This solves the problem that docker.image(...).inside in Jenkinsfile
only passes on the user default group to the running container,
which does not allow us to run docker commands.

This mimics the behaviour in the old slaves that uses the wrapper
image, where the docker user is having gid 1000 in the wrapper, and
the jenkins user is having gid 1000 in the slave.

Related to #12